+++ src/JsonRpc/Exception/InvalidRequestException.php
@@ -10,6 +10,6 @@ class InvalidRequestException extends JsonRpcException
 {
     public function __construct()
     {
-        parent::__construct(JsonRpcError::InvalidRequest->message(), JsonRpcError::InvalidRequest->code());
+        parent::__construct(JsonRpcError::InvalidRequest->message(), (int)JsonRpcError::InvalidRequest->code());
     }
 }
